Implement in-memory audit log queue
This adds an in-memory SQLite-backed audit log queue. The intention is
this can act as a fallback if there is an error constructing a disk
backed queue, for example, if writing to the filesystem causes an error.
A follow-up PR will implement the fallback logic.
