+++ lib/events/auditqueue/auditqueue.go
@@ -46,8 +46,10 @@ var (
 type Kind string
 
 const (
-	// KindSQLite selects the SQLite-backed implementation.
-	KindSQLite Kind = "sqlite"
+	// KindSQLiteDisk selects the SQLite-backed on-disk implementation.
+	KindSQLiteDisk Kind = "sqlite_disk"
+	// KindSQLiteMemory selects the SQLite-backed in-memory implementation.
+	KindSQLiteMemory Kind = "sqlite_memory"
 )
 
 // Config configures a Queue.
@@ -57,7 +59,7 @@ type Config struct {
 	// OrphanScanInterval is how often the queue scans for orphaned audit log
 	// queues.
 	OrphanScanInterval time.Duration
-	// MaxBytes sets the maximum database file size
+	// MaxBytes sets the maximum database size
 	MaxBytes int64
 	// SoftLimit is the size of the audit log queue at which we start logging
 	// warning messages.
@@ -102,8 +104,10 @@ type Queue interface {
 // New constructs a Queue of the given kind.
 func New(kind Kind, cfg Config) (Queue, error) {
 	switch kind {
-	case KindSQLite:
+	case KindSQLiteDisk:
 		return newSQLiteQueue(cfg)
+	case KindSQLiteMemory:
+		return newSQLiteInMemoryQueue(cfg)
 	default:
 		return nil, trace.BadParameter("unknown audit queue kind: '%s'", kind)
 	}
