+++ lib/events/auditqueue/sqlite.go
@@ -26,8 +26,6 @@ import (
 	"log/slog"
 	"net/url"
 	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -36,24 +34,15 @@ import (
 	sqlite "modernc.org/sqlite"
 	sqlite3 "modernc.org/sqlite/lib"
 
-	"github.com/gravitational/teleport"
 	apievents "github.com/gravitational/teleport/api/types/events"
 	"github.com/gravitational/teleport/lib/observability/metrics"
-	"github.com/gravitational/teleport/lib/utils"
 )
 
 const (
 	pollInterval                   = 100 * time.Millisecond
-	incrementalVacuumInterval      = 10 * time.Minute
-	defaultOrphanScanInterval      = 10 * time.Minute
-	queueLockFile                  = "queue.lock"
-	queueDBFile                    = "queue.db"
+	sqlitePageSize                 = 4096 // Bytes
 	auditQueueTable                = "audit_queue"
 	auditDeadLetterTable           = "audit_dead_letter"
-	tmpDirSuffix                   = ".tmp"
-	defaultSoftLimit               = 100 * 1024 * 1024 // 100 MiB
-	softLimitCheckInterval         = time.Minute
-	sqlitePageSize                 = 4096 // Bytes
 	defaultMaxAttempts             = 10
 	defaultDeadLetterSweepInterval = 10 * time.Minute
 	defaultDeadLetterTTL           = 30 * 24 * time.Hour // 30 days
@@ -67,23 +56,9 @@ const (
 	defaultMaxBatch  = 25
 	dequeueBatchSize = 25
 
-	// walJournalSizeLimit is the number of bytes the `-wal` file gets
-	// truncated to in between checkpoints.
-	walJournalSizeLimit = 64 * 1024 * 1024 // 64 MiB
-
 	// busyTimeoutMillis sets the maximum time we will wait for a SQLite DB
 	// operation.
 	busyTimeoutMillis = 5000
-
-	// defaultMaxBytes limits the size of the SQLite database file.
-	defaultMaxBytes int64 = 5 * 1024 * 1024 * 1024 // 5 GiB
-
-	// staleTmpThreshold is how old an in-progress <uuid>.tmp/ directory
-	// must be before the orphan scanner removes it. Under normal circumstances,
-	// A `.tmp` directory will only exist for a few milliseconds.
-	staleTmpThreshold       = time.Hour
-	initQueueDirMaxAttempts = 10
-	initQueueDirRetryDelay  = 50 * time.Millisecond
 )
 
 // A note on `AUTOINCREMENT`, to quote the SQLite docs:
@@ -175,62 +150,20 @@ func bytesToPages(nBytes int64) int64 {
 	return nBytes / sqlitePageSize
 }
 
-func getDSN(dbPath string, maxBytes int64) string {
-	maxPages := bytesToPages(maxBytes)
-	params := url.Values{}
-	params.Add("_pragma", "auto_vacuum(INCREMENTAL)")
-	params.Add("_pragma", "journal_mode(WAL)")
-	params.Add("_pragma", "synchronous(NORMAL)")
-	params.Add("_pragma", fmt.Sprintf("journal_size_limit(%d)", walJournalSizeLimit))
+// addSharedParams adds the shared parameters between both the in-memory sqlite
+// implementation and the on-disk implementation.
+func addSharedParams(params url.Values, maxBytes int64) {
 	params.Add("_pragma", fmt.Sprintf("busy_timeout(%d)", busyTimeoutMillis))
 	params.Add("_pragma", "temp_store(MEMORY)")
-	params.Add("_pragma", fmt.Sprintf("max_page_count(%d)", maxPages))
-	u := url.URL{
-		Scheme:   "file",
-		OmitHost: true,
-		Path:     dbPath,
-		RawQuery: params.Encode(),
-	}
-	return u.String()
+	params.Add("_pragma", fmt.Sprintf("max_page_count(%d)", bytesToPages(maxBytes)))
 }
 
-func newSQLiteQueue(cfg Config) (*sqliteQueue, error) {
-	if cfg.Path == "" {
-		return nil, trace.BadParameter("Path is required to create an sqlite queue")
-	}
-
+// newBaseQueue creates the common core of a sqliteQueue. It is used so
+// that shared initialization between the `sqliteQueue` and the
+// `sqliteInMemoryQueue` can re-use code.
+func newBaseQueue(db *sql.DB, cfg Config) (*sqliteQueue, error) {
 	if err := metrics.RegisterPrometheusCollectors(prometheusCollectors...); err != nil {
 		return nil, trace.Wrap(err)
-	}
-
-	unlock, err := initQueueDir(cfg.Path)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	db, err := initializeDb(cfg.Path, cfg.MaxBytes)
-	if err != nil {
-		_ = unlock()
-		_ = os.RemoveAll(cfg.Path)
-		return nil, trace.Wrap(err)
-	}
-
-	selfStat, err := os.Stat(cfg.Path)
-	if err != nil {
-		db.Close()
-		_ = unlock()
-		_ = os.RemoveAll(cfg.Path)
-		return nil, trace.ConvertSystemError(err)
-	}
-
-	scanInterval := cfg.OrphanScanInterval
-	if scanInterval <= 0 {
-		scanInterval = defaultOrphanScanInterval
-	}
-
-	softLimit := cfg.SoftLimit
-	if softLimit <= 0 {
-		softLimit = defaultSoftLimit
 	}
 
 	maxAttempts := cfg.MaxAttempts
@@ -251,168 +184,19 @@ func newSQLiteQueue(cfg Config) (*sqliteQueue, error) {
 	ctx, cancel := context.WithCancel(context.Background())
 	q := &sqliteQueue{
 		db:                      db,
-		path:                    cfg.Path,
 		toBeWritten:             make(chan writeRequest),
 		maxBatch:                defaultMaxBatch,
 		ctx:                     ctx,
 		cancel:                  cancel,
-		parentDir:               filepath.Dir(cfg.Path),
-		selfStat:                selfStat,
-		unlock:                  unlock,
-		orphanScanInterval:      scanInterval,
-		softLimit:               softLimit,
 		maxAttempts:             maxAttempts,
 		deadLetterSweepInterval: deadLetterSweepInterval,
 		deadLetterTTL:           deadLetterTTL,
 		deadLetterKick:          make(chan struct{}, 1),
 	}
 
-	maxBytes := cfg.MaxBytes
-	if maxBytes <= 0 {
-		maxBytes = defaultMaxBytes
-	}
-	slog.InfoContext(q.ctx, "Audit queue initialized.",
-		"path", cfg.Path,
-		"max_bytes", maxBytes,
-		"soft_limit", softLimit,
-		"max_attempts", maxAttempts,
-		"dead_letter_ttl", deadLetterTTL,
-	)
-
 	q.wg.Go(q.writeLoop)
-	q.wg.Go(q.vacuumLoop)
-	q.wg.Go(q.softLimitLoop)
 
 	return q, nil
-}
-
-func initializeDb(path string, maxBytes int64) (*sql.DB, error) {
-	if maxBytes <= 0 {
-		maxBytes = defaultMaxBytes
-	}
-
-	dbPath := filepath.Join(path, queueDBFile)
-	db, err := sql.Open("sqlite", getDSN(dbPath, maxBytes))
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	db.SetMaxOpenConns(1)
-	if _, err := db.Exec(schemaSQL); err != nil {
-		db.Close()
-		return nil, trace.Wrap(err)
-	}
-	if err := recordTeleportVersion(db, teleport.Version); err != nil {
-		db.Close()
-		return nil, trace.Wrap(err)
-	}
-
-	return db, nil
-}
-
-// initQueueDir creates and initializes the directory where the audit log queue
-// will reside. It does this in a way that is safe from race conditions.
-func initQueueDir(path string) (func() error, error) {
-	tmpPath := path + tmpDirSuffix
-	var err error
-	for attempt := range initQueueDirMaxAttempts {
-		var unlock func() error
-		unlock, err = tryInitQueueDir(path, tmpPath)
-		if err == nil {
-			return unlock, nil
-		}
-
-		isRetryableError := errors.Is(err, utils.ErrUnsuccessfulLockTry) || trace.IsNotFound(err)
-		if !isRetryableError {
-			return nil, trace.Wrap(err)
-		}
-
-		if attempt < initQueueDirMaxAttempts-1 {
-			time.Sleep(initQueueDirRetryDelay)
-		}
-	}
-	return nil, trace.Wrap(err, "failed to initialize audit-queue directory %q after %d attempts (is the system time set correctly?)", path, initQueueDirMaxAttempts)
-}
-
-func tryInitQueueDir(path, tmpPath string) (func() error, error) {
-	if err := os.MkdirAll(tmpPath, 0o700); err != nil {
-		return nil, trace.ConvertSystemError(err)
-	}
-
-	// Take the lock on the `queue.lock` file. This ensures that no other
-	// instances try to adopt this queue.
-	unlock, err := utils.FSTryWriteLock(filepath.Join(tmpPath, queueLockFile))
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	// Remove the `.tmp` suffix, marking this as a live queue.
-	if err := os.Rename(tmpPath, path); err != nil {
-		// Only remove the directory if we own the flock.
-		_ = os.RemoveAll(tmpPath)
-		_ = unlock()
-		return nil, trace.ConvertSystemError(err)
-	}
-	return unlock, nil
-}
-
-// vacuumLoop periodically cleans up deleted records from the SQLite database
-// file.
-func (q *sqliteQueue) vacuumLoop() {
-	timer := time.NewTimer(incrementalVacuumInterval)
-	defer timer.Stop()
-	for {
-		select {
-		case <-q.ctx.Done():
-			return
-		case <-timer.C:
-			if _, err := q.db.ExecContext(q.ctx, "PRAGMA incremental_vacuum"); err != nil {
-				slog.ErrorContext(q.ctx, "Failed to run incremental_vacuum.", "error", err)
-			}
-			timer.Reset(incrementalVacuumInterval)
-		}
-	}
-}
-
-// softLimitLoop periodically stats queue.db and emits a warning when its
-// size exceeds the configured soft limit.
-func (q *sqliteQueue) softLimitLoop() {
-	ticker := time.NewTicker(softLimitCheckInterval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-q.ctx.Done():
-			return
-		case <-ticker.C:
-		}
-		above, size, err := q.aboveSoftLimit()
-		if err != nil {
-			slog.ErrorContext(q.ctx,
-				"Failed to stat audit queue file.",
-				"path", q.path,
-				"error", err,
-			)
-			continue
-		}
-		if above {
-			slog.WarnContext(q.ctx,
-				"audit event queue above soft limit",
-				"path", q.path,
-				"size_bytes", size,
-				"soft_limit_bytes", q.softLimit,
-			)
-			softLimitWarnings.Inc()
-		}
-	}
-}
-
-func (q *sqliteQueue) aboveSoftLimit() (bool, int64, error) {
-	info, err := os.Stat(filepath.Join(q.path, queueDBFile))
-	if err != nil {
-		return false, 0, trace.ConvertSystemError(err)
-	}
-	size := info.Size()
-	return size > q.softLimit, size, nil
 }
 
 // Enqueue writes the event to the SQLite based queue. If you get a nil error
@@ -555,26 +339,7 @@ func isSQLiteFullError(err error) bool {
 	return errors.As(err, &e) && (e.Code() == sqlite3.SQLITE_FULL)
 }
 
-// Run drains the queue. `handler` is the function called for each audit log
-// event that is held within the queue. The audit log queue follows a single
-// consumer model in order to batch events together and commit them as groups.
-func (q *sqliteQueue) Run(ctx context.Context, handler Handler) error {
-	// Ensure we only have a single consumer running.
-	if !q.runMu.TryLock() {
-		return trace.Wrap(ErrAlreadyRunning)
-	}
-	defer q.runMu.Unlock()
-
-	// Startup the orphan scanner and dead-letter sweeper.
-	var wg sync.WaitGroup
-	wg.Go(func() {
-		q.orphanScanLoop(ctx)
-	})
-	wg.Go(func() {
-		q.deadLetterSweepLoop(ctx, handler)
-	})
-	defer wg.Wait()
-
+func (q *sqliteQueue) runPollLoop(ctx context.Context, handler Handler) error {
 	pollTimer := time.NewTimer(pollInterval)
 	defer pollTimer.Stop()
 
@@ -942,384 +707,6 @@ func (q *sqliteQueue) expireDeadLetter() {
 	}
 }
 
-func (q *sqliteQueue) orphanScanLoop(ctx context.Context) {
-	ticker := time.NewTicker(q.orphanScanInterval)
-	defer ticker.Stop()
-	for {
-		q.sweepStaleTmp()
-		q.adoptOrphans(ctx)
-
-		select {
-		case <-ctx.Done():
-			return
-		case <-q.ctx.Done():
-			return
-		case <-ticker.C:
-		}
-	}
-}
-
-// sweepStaleTmp cleans up any audit queue directories that have been orphaned
-// during their initialization phase. This should be rare, but we still want to
-// account for it.
-func (q *sqliteQueue) sweepStaleTmp() {
-	entries, err := os.ReadDir(q.parentDir)
-	if err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx,
-			"Failed to list audit-queue parent directory.",
-			"parent_dir", q.parentDir,
-			"error", err,
-		)
-		return
-	}
-	cutoff := time.Now().Add(-staleTmpThreshold)
-	for _, dirEntry := range entries {
-		if !dirEntry.IsDir() || !strings.HasSuffix(dirEntry.Name(), tmpDirSuffix) {
-			continue
-		}
-		info, err := dirEntry.Info()
-		if err != nil {
-			orphanScanErrors.Inc()
-			slog.ErrorContext(q.ctx,
-				"Failed to stat audit-queue tmp directory.",
-				"name", dirEntry.Name(),
-				"error", err,
-			)
-			continue
-		}
-		if info.ModTime().After(cutoff) {
-			continue
-		}
-
-		// If we get to this point, then the directory is older than the cutoff,
-		// which should not be the case for a tmp directory. These directories
-		// should have their `*.tmp` suffix removed as soon as they are done
-		// initializing, which should be a very quick process.
-		stalePath := filepath.Join(q.parentDir, dirEntry.Name())
-		tryRemoveStaleTmp(q.ctx, stalePath)
-	}
-}
-
-func tryRemoveStaleTmp(ctx context.Context, stalePath string) {
-	unlock, err := utils.FSTryWriteLock(filepath.Join(stalePath, queueLockFile))
-	if err != nil {
-		// The lock is held, so a creator is still active. Leave the directory
-		// for its owner and try again on the next sweep.
-		if errors.Is(err, utils.ErrUnsuccessfulLockTry) {
-			slog.WarnContext(ctx,
-				"Audit-queue tmp directory is still locked past the stale threshold. Leaving it for its owner.",
-				"path", stalePath,
-				"stale_threshold", staleTmpThreshold,
-			)
-			return
-		}
-		orphanScanErrors.Inc()
-		slog.ErrorContext(ctx,
-			"Failed to lock stale audit-queue tmp directory.",
-			"path", stalePath,
-			"error", err,
-		)
-		return
-	}
-	defer func() {
-		if err := unlock(); err != nil {
-			slog.ErrorContext(ctx,
-				"Failed to release stale audit-queue tmp flock.",
-				"path", stalePath,
-				"error", err,
-			)
-		}
-	}()
-
-	// We hold the lock, so the directory is unowned. Tmp directories were
-	// orphaned before they finished initializing, therefore they will have no
-	// audit log events, hence they are safe to remove.
-	if err := os.RemoveAll(stalePath); err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(ctx,
-			"Failed to remove stale audit-queue tmp directory.",
-			"path", stalePath,
-			"error", err,
-		)
-	}
-}
-
-func (q *sqliteQueue) adoptOrphans(ctx context.Context) {
-	entries, err := os.ReadDir(q.parentDir)
-	if err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx,
-			"Failed to list audit-queue parent directory.",
-			"parent_dir", q.parentDir,
-			"error", err,
-		)
-		return
-	}
-	for _, dirEntry := range entries {
-		if ctx.Err() != nil {
-			return
-		}
-		if !dirEntry.IsDir() || strings.HasSuffix(dirEntry.Name(), tmpDirSuffix) {
-			continue
-		}
-
-		entryPath := filepath.Join(q.parentDir, dirEntry.Name())
-		entryStat, err := os.Stat(entryPath)
-		if err != nil {
-			orphanScanErrors.Inc()
-			slog.ErrorContext(q.ctx,
-				"Failed to stat audit-queue candidate directory.",
-				"path", entryPath,
-				"error", err,
-			)
-			continue
-		}
-		if os.SameFile(entryStat, q.selfStat) {
-			continue
-		}
-
-		// If we got to this point, then we have found a directory that is not a
-		// tmp directory and is not the same directory as the one this process
-		// is already using. We are safe to attempt to adopt it.
-		q.tryAdoptOrphan(ctx, entryPath)
-	}
-}
-
-func (q *sqliteQueue) tryAdoptOrphan(ctx context.Context, path string) {
-	unlock, err := utils.FSTryWriteLock(filepath.Join(path, queueLockFile))
-	if err != nil {
-		// This error indicates that the lock has already been taken, hence this
-		// queue is not an orphan. We can skip it.
-		if errors.Is(err, utils.ErrUnsuccessfulLockTry) {
-			return
-		}
-
-		// Otherwise, an error occurred while attempting to take the file lock.
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx,
-			"Failed to attempt orphan flock.",
-			"path", path,
-			"error", err,
-		)
-		return
-	}
-	defer func() {
-		if err := unlock(); err != nil {
-			slog.ErrorContext(q.ctx,
-				"Failed to release orphan flock.",
-				"path", path,
-				"error", err,
-			)
-		}
-	}()
-
-	dbFilePath := filepath.Join(path, queueDBFile)
-	db, err := sql.Open("sqlite", getDSN(dbFilePath, defaultMaxBytes))
-	if err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx, "Failed to open orphan SQLite database.", "path", path, "error", err)
-		return
-	}
-	db.SetMaxOpenConns(1)
-
-	if err := db.PingContext(ctx); err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx,
-			"Failed to connect to orphan SQLite database.",
-			"path", path,
-			"error", err,
-		)
-		_ = db.Close()
-		return
-	}
-
-	migrateErr := q.migrateOrphanDB(ctx, db, filepath.Base(path))
-	if err := db.Close(); err != nil {
-		slog.ErrorContext(q.ctx,
-			"Failed to close orphan SQLite database.",
-			"path", path,
-			"error", err,
-		)
-		return
-	}
-
-	if migrateErr != nil {
-		// If we fail to migrate, we will try again on the next orphan adoption
-		// cycle. Cancellation is a normal shutdown, not a migration failure.
-		if ctx.Err() == nil && q.ctx.Err() == nil {
-			orphanScanErrors.Inc()
-			slog.ErrorContext(q.ctx,
-				"Failed to migrate orphaned audit-queue database.",
-				"path", path,
-				"error", migrateErr,
-			)
-		}
-		return
-	}
-
-	// If we got here, then we have successfully migrated the orphan. We can now
-	// safely remove it.
-	if err := os.RemoveAll(path); err != nil {
-		orphanScanErrors.Inc()
-		slog.ErrorContext(q.ctx,
-			"Failed to remove migrated orphan directory.",
-			"path", path,
-			"error", err,
-		)
-		return
-	}
-	q.clearOrphanWatermarks(ctx, filepath.Base(path))
-	orphansAdopted.Inc()
-	slog.InfoContext(q.ctx, "Adopted orphaned audit-queue directory.", "path", path)
-}
-
-func (q *sqliteQueue) migrateOrphanDB(ctx context.Context, db *sql.DB, name string) error {
-	if err := q.migrateOrphanQueue(ctx, db, name); err != nil {
-		return trace.Wrap(err)
-	}
-	return trace.Wrap(q.migrateOrphanDeadLetter(ctx, db, name))
-}
-
-func (q *sqliteQueue) migrateOrphanQueue(ctx context.Context, orphan *sql.DB, name string) error {
-	return q.migrateOrphanTable(ctx, orphan, name, auditQueueTable,
-		"SELECT id, payload, attempts FROM audit_queue WHERE id > ? ORDER BY id ASC LIMIT ?",
-		"INSERT INTO audit_queue (payload, attempts) VALUES (?, ?)",
-	)
-}
-
-func (q *sqliteQueue) migrateOrphanDeadLetter(ctx context.Context, orphan *sql.DB, name string) error {
-	return q.migrateOrphanTable(ctx, orphan, name, auditDeadLetterTable,
-		"SELECT id, payload, failed_at FROM audit_dead_letter WHERE id > ? ORDER BY id ASC LIMIT ?",
-		"INSERT INTO audit_dead_letter (payload, failed_at) VALUES (?, ?)",
-	)
-}
-
-func orphanWatermarkKey(name, table string) string {
-	return "orphan_migration:" + name + ":" + table
-}
-
-func (q *sqliteQueue) migrateOrphanTable(ctx context.Context, orphan *sql.DB, name, table, selectSQL, insertSQL string) error {
-	watermarkKey := orphanWatermarkKey(name, table)
-	watermark, err := q.readOrphanWatermark(ctx, watermarkKey)
-	if err != nil {
-		return trace.Wrap(err, "reading orphan %s watermark", table)
-	}
-	for {
-		if err := ctx.Err(); err != nil {
-			return trace.Wrap(err)
-		}
-		if err := q.ctx.Err(); err != nil {
-			return trace.Wrap(err)
-		}
-		batch, err := fetchOrphanRows(ctx, orphan, selectSQL, watermark, dequeueBatchSize)
-		if err != nil {
-			return trace.Wrap(err, "fetching orphan %s rows", table)
-		}
-		if len(batch) == 0 {
-			return nil
-		}
-
-		maxID := batch[len(batch)-1].id
-		if err := q.insertMigratedBatch(ctx, insertSQL, batch, watermarkKey, maxID); err != nil {
-			return trace.Wrap(err, "migrating orphan %s rows", table)
-		}
-		watermark = maxID
-
-		ids := make([]int64, len(batch))
-		for i, r := range batch {
-			ids[i] = r.id
-		}
-		if err := deleteIDsFromTable(ctx, orphan, table, ids); err != nil {
-			return trace.Wrap(err, "deleting migrated orphan %s rows", table)
-		}
-	}
-}
-
-func (q *sqliteQueue) readOrphanWatermark(ctx context.Context, key string) (int64, error) {
-	var value string
-	err := q.db.QueryRowContext(ctx,
-		"SELECT value FROM teleport_info WHERE key = ?", key).Scan(&value)
-	if errors.Is(err, sql.ErrNoRows) {
-		return 0, nil
-	}
-	if err != nil {
-		return 0, trace.Wrap(err)
-	}
-	watermark, err := strconv.ParseInt(value, 10, 64)
-	return watermark, trace.Wrap(err)
-}
-
-func (q *sqliteQueue) clearOrphanWatermarks(ctx context.Context, name string) {
-	if _, err := q.db.ExecContext(ctx,
-		"DELETE FROM teleport_info WHERE key IN (?, ?)",
-		orphanWatermarkKey(name, auditQueueTable),
-		orphanWatermarkKey(name, auditDeadLetterTable),
-	); err != nil {
-		slog.ErrorContext(q.ctx,
-			"Failed to clear orphan migration watermarks.",
-			"orphan", name,
-			"error", err,
-		)
-	}
-}
-
-type migratedRow struct {
-	id     int64
-	values []any
-}
-
-func fetchOrphanRows(ctx context.Context, db *sql.DB, selectSQL string, afterID int64, limit int) ([]migratedRow, error) {
-	rows, err := db.QueryContext(ctx, selectSQL, afterID, limit)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	defer rows.Close()
-	cols, err := rows.Columns()
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	var out []migratedRow
-	for rows.Next() {
-		r := migratedRow{values: make([]any, len(cols)-1)}
-		targets := make([]any, 0, len(cols))
-		targets = append(targets, &r.id)
-		for i := range r.values {
-			targets = append(targets, &r.values[i])
-		}
-		if err := rows.Scan(targets...); err != nil {
-			return nil, trace.Wrap(err)
-		}
-		out = append(out, r)
-	}
-	return out, trace.Wrap(rows.Err())
-}
-
-func (q *sqliteQueue) insertMigratedBatch(ctx context.Context, insertSQL string, batch []migratedRow, watermarkKey string, maxID int64) error {
-	tx, err := q.db.BeginTx(ctx, nil)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	defer tx.Rollback()
-	stmt, err := tx.PrepareContext(ctx, insertSQL)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	defer stmt.Close()
-	for _, r := range batch {
-		if _, err := stmt.ExecContext(ctx, r.values...); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-	if _, err := tx.ExecContext(ctx,
-		"INSERT OR REPLACE INTO teleport_info (key, value) VALUES (?, ?)",
-		watermarkKey, strconv.FormatInt(maxID, 10),
-	); err != nil {
-		return trace.Wrap(err)
-	}
-	return trace.Wrap(tx.Commit())
-}
-
 func (q *sqliteQueue) fetch(limit int) ([]Item, error) {
 	return fetchDB(q.ctx, q.db, limit)
 }
@@ -1407,62 +794,6 @@ func deleteByIDs(ctx context.Context, db *sql.DB, table string, items []Item) er
 // ackDB deletes the rows for items from the audit_queue table.
 func ackDB(ctx context.Context, db *sql.DB, items []Item) error {
 	return deleteByIDs(ctx, db, auditQueueTable, items)
-}
-
-func (q *sqliteQueue) Close() error {
-	var errs []error
-	q.closeOnce.Do(func() {
-		q.cancel()
-		q.wg.Wait()
-
-		// Flush the WAL file.
-		if _, err := q.db.Exec("PRAGMA wal_checkpoint(TRUNCATE)"); err != nil {
-			slog.ErrorContext(q.ctx,
-				"Failed to checkpoint WAL on close.",
-				"path", q.path,
-				"error", err,
-			)
-		}
-
-		empty, err := isQueueEmpty(q.db)
-		if err != nil {
-			slog.ErrorContext(q.ctx,
-				"Failed to check whether audit queue is empty on close.",
-				"path", q.path,
-				"error", err,
-			)
-		}
-
-		if err := q.db.Close(); err != nil {
-			errs = append(errs, err)
-		}
-
-		// Remove the directory before releasing the lock. This ensures no other
-		// process can adopt and start draining a queue we are about to delete.
-		if empty {
-			if err := os.RemoveAll(q.path); err != nil {
-				slog.ErrorContext(q.ctx, "Failed to remove empty audit-queue directory on close.", "path", q.path, "error", err)
-			}
-		}
-
-		if q.unlock != nil {
-			if err := q.unlock(); err != nil {
-				errs = append(errs, err)
-			}
-		}
-	})
-	return trace.NewAggregate(errs...)
-}
-
-func isQueueEmpty(db *sql.DB) (bool, error) {
-	var hasRows int
-	err := db.QueryRow(
-		"SELECT EXISTS(SELECT 1 FROM audit_queue) OR EXISTS(SELECT 1 FROM audit_dead_letter)",
-	).Scan(&hasRows)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-	return hasRows == 0, nil
 }
 
 func recordTeleportVersion(db *sql.DB, version string) error {
